Added a lot of documentation.
Added NewWithCapacity to provide capacity hints to avoid re-sizing the maps if the user knows what the size is likely to be.

Refactored the mutating or setter methods to make use of the NewWithCapacity to avoid editing the map directly as much as possible.

Renamed Equal to IsEqual to match the getter syntax